+++ src/homework/Competititors.java
@@ -1,3 +1,5 @@
+package homework;
+
 public interface Competititors {
     int run();
     int jump();
