+++ src/homework/Obstruction.java
@@ -1,3 +1,5 @@
+package homework;
+
 public interface Obstruction {
     String name();
 }
